pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "skin_disease_recognition"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Łukasz Jarzęcki" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "albumentations>=2.0.8",
    "dvc>=3.66.1",
    "fastapi>=0.128.7",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "kagglehub>=0.4.2",
    "mlflow (>=3.7.0,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pip",
    "pre-commit>=4.5.1",
    "protobuf~=6.0",
    "pytest",
    "python-dotenv",
    "python-multipart>=0.0.22",
    "ruff",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torchmetrics>=1.8.2",
    "uvicorn>=0.40.0",
]
requires-python = "~=3.13.0"

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.ruff]
line-length = 88
indent-width = 4
src = ["src"]
include = ["pyproject.toml", "src/skin_disease_recognition/**/*.py", "notebooks/*"]

[tool.ruff.lint]
# Select which rule categories are enabled
# (E: pycodestyle errors, F: pyflakes, I: import-related, B: bugbears, UP: plugins)
select = ["E", "F", "I", "UP", "B"]
fixable = ["ALL"]

[tool.ty.rules]


[tool.ruff.lint.isort]
known-first-party = ["skin_disease_recognition"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "single"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
pytorch = [
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

